+++ LEAF_Agent/routeConditionalData.go
@@ -1,0 +1,50 @@
+package main
+
+import (
+	"github.com/department-of-veterans-affairs/LEAF/pkg/form/query"
+)
+
+type RouteConditionalDataPayload struct {
+	ActionType string      `json:"actionType"`
+	Query      query.Query `json:"query"`
+}
+
+func routeConditionalData(task Task, payload RouteConditionalDataPayload) error {
+	// Initialize query. At minimum it should only return records that match the stepID
+	query := query.Query{
+		Terms: []query.Term{
+			{
+				ID:       "stepID",
+				Operator: "=",
+				Match:    task.StepID,
+			},
+		},
+	}
+
+	// Only use allowed terms in the query
+	for _, term := range payload.Query.Terms {
+		switch term.ID {
+		case "data",
+			"serviceID",
+			"title",
+			"userID",
+			"dateInitiated",
+			"dateSubmitted",
+			"categoryID",
+			"dependencyID",
+			"stepAction":
+			query.Terms = append(query.Terms, term)
+		}
+	}
+
+	records, err := FormQuery(task.SiteURL, query, "&x-filterData=")
+	if err != nil {
+		return err
+	}
+
+	for recordID := range records {
+		TakeAction(task.SiteURL, recordID, task.StepID, payload.ActionType, "")
+	}
+
+	return nil
+}
